automatically read memories when generating the system message
