+++ app/api/chat/route.ts
@@ -16,7 +16,7 @@ import {
   FunctionName,
   availableFunctions,
 } from "./functions";
-
+import { getMemory } from "@/app/utils/github";
 export const runtime = "edge";
 const MODEL = process.env.MODEL_VERSION || "gpt-4-1106-preview";
 
@@ -38,9 +38,9 @@ function signatureFromArgs(args: Record<string, unknown>) {
     .join(", ");
 }
 
-function getSystemMessage(
+async function getSystemMessage(
   customInstructions: string,
-): ChatCompletionMessageParam {
+): Promise<ChatCompletionMessageParam> {
   const instructions = `You are a helpful coding assistant that assists users with coding questions.
 
   * You have been provided a number of functions that load data from GitHub.com. 
@@ -51,7 +51,10 @@ function getSystemMessage(
   * Be concise and helpful in your responses.
   * Most users are developers. Use technical terms freely and avoid over simplification.
   * If the user asks about your capabilities, please respond with a summary based on the list of functions provided to you. Don't worry too much about specific functions, instead give them an overview of how you can use these functions to help the user.  
-  * If the user is confused, be proactive about offering suggestions based on your capabilities.${
+  * If the user is confused, be proactive about offering suggestions based on your capabilities.
+  When you speak to users, you have the ability to record memories about the person and their preferences. Here are the memories you have previously recorded for the current user. Use these memories to improve the users experience:
+  ${await getMemory()}
+  ${
     customInstructions
       ? ` The user has provided the following additional instructions:${customInstructions}`
       : ``
@@ -104,7 +107,7 @@ export async function POST(req: Request) {
   } = body;
 
   const data = new experimental_StreamData();
-  const systemMessage = getSystemMessage(settings.customInstructions);
+  const systemMessage = await getSystemMessage(settings.customInstructions);
 
   if (imageUrl) {
     return handleImageMessage(imageUrl, messages, systemMessage);
